bugfix(codec): Fixed the import error. The test is running fine
